Fixed update/insert issue on all top-level entities
git-svn-id: https://code.elgg.org/elgg/trunk@912 36083f99-b078-4883-b0ff-0f9b5a30f544

+++ engine/lib/sites.php
@@ -269,15 +269,17 @@
 		if ($row)
 		{
 			// Exists and you have access to it
-			$result = update_data("UPDATE {$CONFIG->dbprefix}sites_entity set name='$name', description='$description', url='$url' where guid=$guid");
-			if ($result!=false)
-			{
-				// Update succeeded, continue
-				$entity = get_entity($guid);
-				if (trigger_elgg_event('update',$entity->type,$entity)) {
-					return true;
-				} else {
-					delete_entity($guid);
+			if ($exists = get_data_row("select guid from {$CONFIG->dbprefix}sites_entity where guid = {$guid}")) {
+				$result = update_data("UPDATE {$CONFIG->dbprefix}sites_entity set name='$name', description='$description', url='$url' where guid=$guid");
+				if ($result!=false)
+				{
+					// Update succeeded, continue
+					$entity = get_entity($guid);
+					if (trigger_elgg_event('update',$entity->type,$entity)) {
+						return true;
+					} else {
+						delete_entity($guid);
+					}
 				}
 			}
 			else
